+++ src/main/clojure/clojure/tools/analyzer/jvm.clj
@@ -34,6 +34,7 @@
              [box :refer [box]]
              [constant-lifter :refer [constant-lift]]
              [classify-invoke :refer [classify-invoke]]
+             [process-method-value :refer [process-method-value]]
              [validate :refer [validate]]
              [infer-tag :refer [infer-tag]]
              [validate-loop-locals :refer [validate-loop-locals]]
@@ -90,8 +91,7 @@
     #'clojure.core/when-not
     #'clojure.core/while
     #'clojure.core/with-open
-    #'clojure.core/with-out-str
-    })
+    #'clojure.core/with-out-str})
 
 (def specials
   "Set of the special forms for clojure in the JVM"
@@ -127,13 +127,31 @@
   (let [sym-ns (namespace form)]
     (if-let [target (and sym-ns
                          (not (resolve-ns (symbol sym-ns) env))
-                         (maybe-class-literal sym-ns))]          ;; Class/field
-      (let [opname (name form)]
-        (if (and (= (count opname) 1)
-                 (Character/isDigit (char (first opname))))
-          form ;; Array/<n>
-          (with-meta (list '. target (symbol (str "-" opname))) ;; transform to (. Class -field)
-            (meta form))))
+                         (maybe-class-literal sym-ns))]
+      (let [opname (name form)
+            opsym (symbol opname)]
+        (cond
+          ;; Array/<n>, leave as is
+          (and (= (count opname) 1)
+               (Character/isDigit (char (first opname))))
+          form
+
+          ;; Class/.method or Class/new, leave as is to be parsed as :maybe-host-form -> :method-value
+          (or (.startsWith ^String opname ".")
+              (= "new" opname))
+          form
+
+          ;; Class/name where name is a static field,  desugar to (. Class -name) as before
+          ;; But if :param-tags are present and methods with the same name exist, then leave as is to go through
+          ;; :method-value path
+          (static-field target opsym)
+          (if (and (param-tags-of form)
+                   (seq (filter :return-type (static-members target opsym))))
+            form
+            (with-meta (list '. target (symbol (str "-" opname)))
+              (meta form)))
+
+          :else form))
       form)))
 
 (defn desugar-host-expr [form env]
@@ -143,13 +161,23 @@
             opns   (namespace op)]
         (if-let [target (and opns
                              (not (resolve-ns (symbol opns) env))
-                             (maybe-class-literal opns))] ; (class/field ..)
+                             (maybe-class-literal opns))]
 
-          (let [op (symbol opname)]
-            (with-meta (list '. target (if (zero? (count expr))
-                                         op
-                                         (list* op expr)))
-              (meta form)))
+          (cond
+           ;; (Class/new args), (Class/.method target args), (^[pt] Class/method args)
+           ;; -> leave as-is, will be analyzed as invoke of method-value
+           (or (= "new" opname)
+               (.startsWith ^String opname ".")
+               (param-tags-of op))
+           form
+
+           ;; (Class/method args) -> (. Class (method args))
+           :else
+           (let [op-sym (symbol opname)]
+             (with-meta (list '. target (if (seq expr)
+                                          (list* op-sym expr)
+                                          op-sym))
+               (meta form))))
 
           (cond
            (.startsWith opname ".")     ; (.foo bar ..)
@@ -456,6 +484,7 @@
     #'box
 
     #'analyze-host-expr
+    #'process-method-value
     #'validate-loop-locals
     #'validate
     #'infer-tag